Swift: Remove redundant APIs
